services/ML-service/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
huggingface_hub
timm